requirements.txt
pywin32
fastapi
uvicorn
cryptography
pyside6
pillow